+++ src/cobib/utils/logging.py
@@ -21,6 +21,8 @@ from .rel_path import RelPath
 # NOTE: we add a custom HINT level which has a higher priority than WARNING and, thus, can be used
 # to provide information to the user that might be useful to see at runtime.
 logging.addLevelName(35, "HINT")
+# NOTE: we also add a custom level DEPRECATED which has an even higher priority.
+logging.addLevelName(45, "DEPRECATED")
 
 
 class _StderrHandler(logging.StreamHandler):  # type: ignore[type-arg]